requirements.txt
Click==7.0
Flask==1.1.1
Flask-PyMongo==2.3.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pymongo==3.10.1
python-dotenv==0.12.0
Werkzeug==0.16.1
